Tell Emacs how to handle this file
This file uses iso-8859-1 encoding, tabs instead of spaces, and single spaces
at the end of the line.

Most Emacs configurations are likely to change the encoding to UTF-8, the tabs
to spaces, and remove the single spaces at the end of the line.

This changes forbids most Emacs configurations from doing so.